src/shared/requirements.txt
# Dilux Database Backup - Shared Dependencies

# Azure SDK
azure-storage-blob>=12.19.0
azure-storage-queue>=12.9.0
azure-data-tables>=12.5.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# Data validation
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Logging
structlog>=24.1.0
